refactor TextfieldShowcaseActivity
